libnativehelper is available to com.android.bluetooth.updatable
The library is currently depended on by Bluetooth.apk in the bluetooth
apex. Until the dependency is dropped, allow libnativehelper to be
included in the apex.

Or, we might want to allow libnativehelper to be available to
everywhere. This makes sense because the library actually doesn't share
any private info with ART. It is a thin wrapper around the standard JNI
to ease the creation of jni libraries.

Exempt-From-Owner-Approval: cherry-pick from internal

Bug: 143733063
Bug: 147671264
Test: m

Merged-In: Id7b02ebd248be9a5ec644053bd06e510fd975727
(cherry picked from commit 84654d72e2ae33d94fb44699c22695d379ddfb56)
Change-Id: Id7b02ebd248be9a5ec644053bd06e510fd975727