libnativehelper: Add libnativehelper_lazy
Add lazy loading support to libnativehelper.

Framework libraries depend on libnativehelper.so. There is some code,
such as bootanimation, that depend on framework libraries, and run
before the Android Runtime has started. Now that libnativehelper.so is
in the ART module, libnativehelper.so may not be available when
bootanimation runs.

Bug: 179115178
Test: atest MtsLibnativehelperLazyTestCases
Change-Id: I2cabde9205599bb657be7cad92eb9c103ace41fc