Add jniThrowErrnoException in libnativehelper
Since a few places in frameworks and libcore implement their own
functions to throw android.system.ErrnoException, these have been
integrated in libnativehelper to avoid duplication.

As part of the change a more generic ThrowException function is
provided that can take a variable number of arguments, rather
than just a description string. Due to this, the code now
constructs the exception object itself and then passes it to the
JNIEnv.

A helper macro is also added to convert the C string to jstring
for the description parameter, which is usually the first
parameter in the exception constructors.

Bug: 180958753
Fix: 180958753
Test: atest MtsLibnativehelperTestCases
Test: atest MtsLibnativehelperLazyTestCases
Change-Id: I4841cf9924a48eaf0b6965c203c9ebe851e443ff